@@ -5,11 +5,11 @@ description = ""
 authors = ["Evan Thomas <evan@evanjt.com>"]
 
 [tool.poetry.dependencies]
-python = "^3.9"
-fastapi = "^0.78.0"
-uvicorn = "^0.18.1"
-SQLAlchemy = "^1.4.36"
-pydantic = "^1.9.1"
+python = "^3.10"
+fastapi = "^0.85.1"
+uvicorn = "^0.19.0"
+SQLAlchemy = "^1.4.41"
+pydantic = "^1.10.2"
 email-validator = "^1.2.1"
 psycopg2 = "^2.9.3"
 SQLAlchemy-Utils = "^0.38.2"
@@ -25,11 +25,6 @@ requests = "^2.27.1"
 pytest-watch = "^4.2.0"
 ptw = "^1.0.1"
 jupyter = "^1.0.0"
-moto = "^4.0.8"
-aiobotocore = "^2.4.0"
-aioboto3 = "^10.1.0"
-pytest-aiomoto = "^0.6.1"
-s3fs = "^2022.10.0"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
